requirements.txt
paddlepaddle-gpu
torch
torchvision
torchaudio
transformers
paddleocr
numpy
pandas
Pillow
sklearn
dotenv
openai